Refactor test_gates
